// design/shared/types/spacingTypes.ts
type SpacingScale = 0 | 0.5 | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10;

export type SpacingStyles = {
  padding?: SpacingScale;
  spacing?: SpacingScale;
  spacingTop?: SpacingScale;
  spacingBottom?: SpacingScale;
  spacingLeft?: SpacingScale;
  spacingRight?: SpacingScale;
  spacingVertical?: SpacingScale;
  spacingHorizontal?: SpacingScale;
};